Refactor: chanto atributes from decimal to double in models and derivates

+++ src/Dto/ItemsOrderUserDto.cs
@@ -9,10 +9,10 @@ namespace OrderService.src.Dto
     {
         public string ProductName { get; set; } = string.Empty!;
 
-        public decimal UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
 
         public int Quantity { get; set; }
 
-        public decimal Subtotal { get; set; }
+        public double Subtotal { get; set; }
     }
 }
